LPC84x / System - conversion from CPP to C files and namespace cleanup of some system files
