+++ include/targets/LPC84x/lpc84x_port.hpp
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
 // @file    lpc84x_port.hpp
 // @brief   NXP LPC84x port class.
-// @date    23 May 2018
+// @date    19 June 2018
 // ----------------------------------------------------------------------------
 //
 // Xarmlib 0.1.0 - https://github.com/hparracho/Xarmlib
@@ -33,7 +33,7 @@
 #define __XARMLIB_TARGETS_LPC84X_PORT_HPP
 
 #include "system/target"
-#include "targets/LPC84x/lpc84x_cmsis.hpp"
+#include "targets/LPC84x/lpc84x_cmsis.h"
 #include "targets/LPC84x/lpc84x_pin.hpp"
 
 namespace xarmlib
