+++ include/targets/LPC84x/lpc84x_section_macros.h
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
-// @file    lpc84x_section_macros.hpp
+// @file    lpc84x_section_macros.h
 // @brief   Macros to allow code/data to be placed into different memory banks.
-// @date    18 May 2018
+// @date    19 June 2018
 // ----------------------------------------------------------------------------
 //
 // Xarmlib 0.1.0 - https://github.com/hparracho/Xarmlib
@@ -35,15 +35,8 @@
 // this code.
 // ----------------------------------------------------------------------------
 
-#ifndef __XARMLIB_TARGETS_LPC84X_SECTION_MACROS_HPP
-#define __XARMLIB_TARGETS_LPC84X_SECTION_MACROS_HPP
-
-namespace xarmlib
-{
-namespace targets
-{
-namespace lpc84x
-{
+#ifndef __XARMLIB_TARGETS_LPC84X_SECTION_MACROS_H
+#define __XARMLIB_TARGETS_LPC84X_SECTION_MACROS_H
 
 
 
@@ -117,8 +110,4 @@ namespace lpc84x
 
 
 
-} // namespace lpc84x
-} // namespace targets
-} // namespace xarmlib
-
-#endif // __XARMLIB_TARGETS_LPC84X_SECTION_MACROS_HPP
+#endif // __XARMLIB_TARGETS_LPC84X_SECTION_MACROS_H
