+++ include/targets/LPC84x/lpc84x_syscon_power.hpp
@@ -1,7 +1,7 @@
 // ----------------------------------------------------------------------------
 // @file    lpc84x_syscon_power.hpp
 // @brief   NXP LPC84x SYSCON power control / brown-out classes.
-// @date    18 May 2018
+// @date    19 June 2018
 // ----------------------------------------------------------------------------
 //
 // Xarmlib 0.1.0 - https://github.com/hparracho/Xarmlib
@@ -32,7 +32,7 @@
 #ifndef __XARMLIB_TARGETS_LPC84X_SYSCON_POWER_HPP
 #define __XARMLIB_TARGETS_LPC84X_SYSCON_POWER_HPP
 
-#include "targets/LPC84x/lpc84x_cmsis.hpp"
+#include "targets/LPC84x/lpc84x_cmsis.h"
 
 namespace xarmlib
 {
