+++ source/system/newlib_stubs.c
@@ -1,8 +1,8 @@
 // ----------------------------------------------------------------------------
-// @file    newlib_stubs.cpp
+// @file    newlib_stubs.c
 // @brief   Support files for GNU libc. These functions will replace or
 //          extend some of the newlib functionality.
-// @date    13 June 2018
+// @date    21 June 2018
 // ----------------------------------------------------------------------------
 //
 // Xarmlib 0.1.0 - https://github.com/hparracho/Xarmlib
@@ -35,7 +35,7 @@
 
 
 
-#if !defined(MCUXPRESSO_MANAGED_LINKER_SCRIPTS) && !defined(CPP_NO_HEAP)
+#if !defined (MCUXPRESSO_MANAGED_LINKER_SCRIPTS) && !defined (CPP_NO_HEAP)
 
 #include <sys/types.h>
 #include <errno.h>
@@ -46,7 +46,7 @@
 // Copyright (c) 2016 Liviu Ionescu.
 
 // A custom _sbrk() function to match the settings defined by the linker script.
-extern "C" caddr_t _sbrk(int incr)
+caddr_t _sbrk(int incr)
 {
     extern char __heap_begin;   // Defined in the linker script
     extern char __heap_limit;   // Defined in the linker script
@@ -87,18 +87,19 @@ extern "C" caddr_t _sbrk(int incr)
     return (caddr_t)current_block_address;
 }
 // ----------------------------------------------------------------------------
-#endif // !MCUXPRESSO_MANAGED_LINKER_SCRIPTS
+#endif // !MCUXPRESSO_MANAGED_LINKER_SCRITS && !CPP_NO_HEAP
 
 
 
 
-extern "C" __attribute__ ((weak, noreturn))
+
+__attribute__ ((weak, noreturn))
 void _exit(int code __attribute__ ((unused)))
 {
-#ifndef DEBUG
+#ifdef NDEBUG
     NVIC_SystemReset();
 #else
-    while(true)
+    while(1)
     {}
 #endif
 }
@@ -106,7 +107,7 @@ void _exit(int code __attribute__ ((unused)))
 
 
 
-extern "C" __attribute__ ((weak, noreturn))
+__attribute__ ((weak, noreturn))
 void abort(void)
 {
     _exit(1);
