+++ source/targets/LPC84x/lpc84x_startup_hooks.cpp
@@ -33,9 +33,9 @@
 
 #ifdef __LPC84X__
 
-#include "targets/LPC84x/lpc84x_cmsis.hpp"
+#include <targets/LPC84x/lpc84x_cmsis.h>
 #include "targets/LPC84x/lpc84x_faim.hpp"
-#include "targets/LPC84x/lpc84x_romdivide.hpp"
+#include <targets/LPC84x/lpc84x_romdivide.h>
 #include "targets/LPC84x/lpc84x_swm.hpp"
 #include "targets/LPC84x/lpc84x_syscon_clock.hpp"
 #include "targets/LPC84x/lpc84x_syscon_power.hpp"
